+++ orchestrator/src/formatting_orchestrator.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 
-use crate::{command_line::FormatMode, file_formatter::FileFormatter, predule::ErrHandler};
+use crate::{command_line::FormatMode, file_formatter::FileFormatter, prelude::ErrHandler};
 use log::LevelFilter;
 
 pub trait FormatterConfiguration {
@@ -20,14 +20,14 @@ impl FormattingOrchestrator {
         match config.mode() {
             FormatMode::Check if config.is_stdin() => file_formatter.check_stdin(err_handler),
             FormatMode::Stdout if config.is_stdin() => {
-                file_formatter.format_stdin_to_stdout(err_handler)
+                file_formatter.format_stdin_to_stdout(err_handler);
             }
             mode => match config.get_paths() {
                 Ok(paths) => match mode {
                     FormatMode::Check => file_formatter.check_files(&paths, err_handler),
                     FormatMode::Files => file_formatter.format_files(&paths, err_handler),
                     FormatMode::Stdout => {
-                        file_formatter.format_files_to_stdout(&paths, err_handler)
+                        file_formatter.format_files_to_stdout(&paths, err_handler);
                     }
                 },
                 Err(e) => err_handler(e),