+++ orchestrator/src/lib.rs
@@ -7,7 +7,7 @@ pub mod formatting_orchestrator;
 pub trait ErrHandler: Fn(anyhow::Error) + Sync {}
 impl<T: Fn(anyhow::Error) + Sync> ErrHandler for T {}
 
-pub mod predule {
+pub mod prelude {
     pub use super::*;
     pub use crate::command_line::{pasfmt_config, *};
     pub use crate::file_formatter::*;